+++ internal/impl/mongodb/common.go
@@ -136,6 +136,8 @@ const (
 	OperationUpdateOne Operation = "update-one"
 	// OperationFindOne Find one operation.
 	OperationFindOne Operation = "find-one"
+	// OperationAggregate Execute Aggregation Pipeline operation.
+	OperationAggregate Operation = "aggregate"
 	// OperationInvalid Invalid operation.
 	OperationInvalid Operation = "invalid"
 )
@@ -144,7 +146,8 @@ func (op Operation) isDocumentAllowed() bool {
 	switch op {
 	case OperationInsertOne,
 		OperationReplaceOne,
-		OperationUpdateOne:
+		OperationUpdateOne,
+		OperationAggregate:
 		return true
 	default:
 		return false
@@ -202,6 +205,8 @@ func NewOperation(op string) Operation {
 		return OperationUpdateOne
 	case "find-one":
 		return OperationFindOne
+	case "aggregate":
+		return OperationAggregate
 	default:
 		return OperationInvalid
 	}
@@ -220,6 +225,7 @@ func processorOperationDocs(defaultOperation Operation) *service.ConfigField {
 		string(OperationReplaceOne),
 		string(OperationUpdateOne),
 		string(OperationFindOne),
+		string(OperationAggregate),
 	).Description("The mongodb operation to perform.").
 		Default(string(defaultOperation))
 }
@@ -242,7 +248,7 @@ func operationFromParsed(pConf *service.ParsedConfig) (operation Operation, err 
 	}
 
 	if operation = NewOperation(operationStr); operation == OperationInvalid {
-		err = fmt.Errorf("mongodb operation '%s' unknown: must be insert-one, delete-one, delete-many, replace-one or update-one", operationStr)
+		err = fmt.Errorf("mongodb operation %q unknown: must be insert-one, delete-one, delete-many, replace-one, update-one or aggregate", operationStr)
 	}
 	return
 }
@@ -325,7 +331,7 @@ func writeMapsFields() []*service.ConfigField {
 	return []*service.ConfigField{
 		service.NewBloblangField(commonFieldDocumentMap).
 			Description("A bloblang map representing a document to store within MongoDB, expressed as https://www.mongodb.com/docs/manual/reference/mongodb-extended-json/[extended JSON in canonical form^]. The document map is required for the operations " +
-				"insert-one, replace-one and update-one.").
+				"insert-one, replace-one, update-one and aggregate.").
 			Examples(mapExamples()...).
 			Default(""),
 		service.NewBloblangField(commonFieldFilterMap).